requirements.txt
Flask==3.1.2
gunicorn==23.0.0
Pillow==11.3.0
reportlab==4.4.6